split out tmux commands
