+++ mux/src/lib.rs
@@ -34,6 +34,7 @@ pub mod ssh;
 pub mod tab;
 pub mod termwiztermtab;
 pub mod tmux;
+pub mod tmux_commands;
 pub mod window;
 
 use crate::activity::Activity;
